+++ app/index.tsx
@@ -1,10 +1,14 @@
+import firebaseConfig from "../src/config/firebaseConfig";
+import useFirebase from "../src/hooks/useFirebase";
+//import Router from "../src/screens/Router"; //verificar uso
 import { useFonts } from 'expo-font';
-
 import { StatusBar } from 'expo-status-bar';
-
 import { Main } from '../src/Main';
 
 function App() {
+  const firebaseApp = useFirebase(firebaseConfig);
+  if (firebaseApp == null) return <div>Loading...</div>;
+
   const [isFontsLoaded] = useFonts({
     'GeneralSans-400': require('../assets/fonts/GeneralSans-Regular.otf'),
     'GeneralSans-600': require('../assets/fonts/GeneralSans-Semibold.otf'),
@@ -14,6 +18,8 @@ function App() {
   if (!isFontsLoaded) {
     return null;
   }
+
+  //return <Router />; //verificar uso
 
   return (
     <>
